Adjusted supabase loading mechanism
