+++ src/routes/(connected)/+layout.server.ts
@@ -1,11 +1,42 @@
+import { SUPABASE_ANON_KEY, SUPABASE_URL } from '$env/static/private'
 import type { LayoutServerLoad } from './$types'
+// import type { LayoutLoad } from './$types'
+import { createBrowserClient, createServerClient, isBrowser } from '@supabase/ssr'
 
-export const load: LayoutServerLoad = async ({ locals: { safeGetSession }, cookies }) => {
-  const { session, user } = await safeGetSession()
+export const load: LayoutServerLoad = async ({ locals: { safeGetSession }, fetch, depends, data }) => {
+  const { user } = await safeGetSession()
 
-  return {
-    session,
-    user,
-    cookies: cookies.getAll(),
-  }
+  depends('supabase:auth')
+
+  const supabase = isBrowser()
+    ? createBrowserClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+        global: {
+          fetch,
+        },
+      })
+    : createServerClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+        global: {
+          fetch,
+        },
+        cookies: {
+          getAll() {
+            return data.cookies
+          },
+        },
+      })
+
+  /**
+   * It's fine to use `getSession` here, because on the client, `getSession` is
+   * safe, and on the server, it reads `session` from the `LayoutData`, which
+   * safely checked the session using `safeGetSession`.
+   */
+  const { data: { session }, } = await supabase.auth.getSession()
+
+  return { supabase, session }
+
+  // return {
+  //   session,
+  //   user,
+  //   cookies: cookies.getAll(),
+  // }
 }
