+++ src/routes/(connected)/+layout.ts
@@ -1,35 +1,35 @@
-import { SUPABASE_ANON_KEY, SUPABASE_URL } from '$env/static/private'
+// import { SUPABASE_ANON_KEY, SUPABASE_URL } from '$env/static/private'
 import type { LayoutLoad } from './$types'
-import { createBrowserClient, createServerClient, isBrowser } from '@supabase/ssr'
+// import { createBrowserClient, createServerClient, isBrowser } from '@supabase/ssr'
 
 export const load: LayoutLoad = async ({ fetch, data, depends }) => {
   depends('supabase:auth')
 
-  const supabase = isBrowser()
-    ? createBrowserClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
-        global: {
-          fetch,
-        },
-      })
-    : createServerClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
-        global: {
-          fetch,
-        },
-        cookies: {
-          getAll() {
-            return data.cookies
-          },
-        },
-      })
+//   const supabase = isBrowser()
+//     ? createBrowserClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+//         global: {
+//           fetch,
+//         },
+//       })
+//     : createServerClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+//         global: {
+//           fetch,
+//         },
+//         cookies: {
+//           getAll() {
+//             return data.cookies
+//           },
+//         },
+//       })
 
-  /**
-   * It's fine to use `getSession` here, because on the client, `getSession` is
-   * safe, and on the server, it reads `session` from the `LayoutData`, which
-   * safely checked the session using `safeGetSession`.
-   */
-  const {
-    data: { session },
-  } = await supabase.auth.getSession()
+//   /**
+//    * It's fine to use `getSession` here, because on the client, `getSession` is
+//    * safe, and on the server, it reads `session` from the `LayoutData`, which
+//    * safely checked the session using `safeGetSession`.
+//    */
+//   const {
+//     data: { session },
+//   } = await supabase.auth.getSession()
 
-  return { supabase, session }
+//   return { supabase, session }
 }
